refactor(state): move addToCart function to CartContainer
